MAJOR UPDATE - Menu Page -Menu filled with dummy data using SampleData.cs. -Images added for each Menu Item. -Menu Items grouped by ItemType. -Styling complete for Desktop View (Still need to do Mobile).

+++ MasciApps-Ass2/Models/Item.cs
@@ -11,7 +11,7 @@ namespace MasciApps_Ass2.Models
         public virtual int ItemTypeId { get; set; }
         public virtual int ItemLabelId { get; set; }
         public virtual string Name { get; set; }
-        public virtual double Price { get; set; }
+        public virtual decimal Price { get; set; }
         public virtual string ImageUrl { get; set; }
         public virtual string ShortDescription { get; set; }
         public virtual string LongDescription { get; set; }
